Fixed a typo when checking qa in the pimp command
The first argument to the pimp command is either "q" or "a".
In my conditional, I was checking if the first argument was either not q
OR not a. This could cause a problem if the user typed a as q is tested first

However, now this was changed to AND so if the first argument is not q AND
it's not a, then there was a problem.