Changed XMLHttpRequest and rearranged conditionals in onreadystatechange
Now, XMLHttpRequest will work for earlier versions of IE. And, in the onreadystatechange function, I moved == 200 to a separate inner conditional so you don't have to get .status each time onreadystatechange calls and the readyState is not 4. Now, when the ready state is 4, if status is not 200, then there was an error.